feat(clay-drop-down): create DropDownBasic high-level component
